:recycle: move `soup_to_weekly_taskboards( )` func to \data_formatting.py
I just thought it made more sense to have a separate file for data formatting.
