+++ wtb/app/main.py
@@ -1,8 +1,8 @@
 import toml
 
+from app.utils.data_formatting import soup_to_weekly_taskboards
 from app.utils.internal_dev import print_taskboards
 from app.utils.os_structure import save_weekly_taskboards
-from app.utils.string_process import soup_to_weekly_taskboards
 from app.utils.webscrape import get_soup_from_altiplan
 
 if __name__ == "__main__":
